requirements.txt
qrcode==7.4.2
Pillow==10.4.0
customtkinter==5.2.2
